update user register

+++ model/project/user_announcement_read.go
@@ -5,7 +5,7 @@ import "time"
 // UserAnnouncementRead 用户公告阅读记录
 type UserAnnouncementRead struct {
 	ID             int64      `json:"id" gorm:"primaryKey;autoIncrement;comment:主键ID"`
-	UserID         int64      `json:"user_id" gorm:"not null;comment:用户ID"`
+	UserID         int        `json:"user_id" gorm:"not null;comment:用户ID"`
 	AnnouncementID int64      `json:"announcement_id" gorm:"not null;comment:公告ID"`
 	IsRead         int        `json:"is_read" gorm:"default:0;comment:是否已读"`
 	ReadTime       *time.Time `json:"read_time" gorm:"comment:阅读时间"`
